% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynplot.R
\name{dynplot}
\alias{dynplot}
\title{Create a dynplot with a specified layout}
\usage{
dynplot(dataset, trajectory, layout)
}
\arguments{
\item{dataset}{A dynwrap dataset object, typically containing a trajectory}

\item{trajectory}{The trajectory dynwrap object, if available}

\item{layout}{A \verb{layout_*} function from dynplot, such as \code{\link[=layout_dimred]{layout_dimred()}} or \code{\link[=layout_graph]{layout_graph()}}}
}
\value{
A ggplot2 object, with the processed data in \code{plot$data} and \code{attr(plot$data, "data")}
}
\description{
Create a dynplot with a specified layout
}
\examples{
library(ggplot2)
dynplot_dimred(example_bifurcating) +
  geom_cell_point(aes(colour = select_feature_expression("G1", d = .data))) +
  scale_expression_colour() +
  new_scale_colour() +
  geom_trajectory_segments(aes(colour = edge_id))

}
