% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynplot_dendro.R
\name{dynplot_dendro}
\alias{dynplot_dendro}
\title{Plot a trajectory as a dendrogram}
\usage{
dynplot_dendro(
  dataset,
  trajectory = dataset,
  diag_offset = 0.05,
  y_offset = 0.2
)
}
\arguments{
\item{dataset}{A dynwrap dataset object, typically containing a trajectory.}

\item{trajectory}{The trajectory dynwrap object, if available.}

\item{diag_offset}{The x-offset (percentage of the edge lengths) between milestones}

\item{y_offset}{The y-offset (percentage of the edge lengths) between milestones}
}
\value{
A ggplot2 object, with the processed data in \code{plot$data} and \code{attr(plot$data, "data")}.
}
\description{
Plot a trajectory as a dendrogram
}
\examples{
library(ggplot2)
data(example_bifurcating)
dataset <- example_bifurcating

dynplot_dendro(dataset) +
  geom_trajectory_segments() +
  geom_trajectory_connection() +
  geom_cell_point(size = 2, colour = "black") +
  geom_cell_point(aes(colour = milestone_percentages), size = 1.8) +
  geom_milestone_label(aes(fill = milestone_id)) +
  scale_milestones_fill() +
  scale_milestones_colour()

dynplot_dendro(dataset) +
  geom_trajectory_segments(aes(color = milestone_percentages), size = 5, color = "#333333") +
  geom_trajectory_connection() +
  scale_milestones_colour() +
  new_scale_colour() +
  geom_cell_point(aes(colour = select_feature_expression("G1", d = .data))) +
  scale_expression_colour() +
  new_scale_colour() +
  geom_milestone_label(aes(fill = milestone_id)) +
  scale_milestones_fill()
}
\seealso{
\code{\link[=dynplot_dendro]{dynplot_dendro()}}, \code{\link[=dynplot_dimred]{dynplot_dimred()}}, \code{\link[=dynplot_graph]{dynplot_graph()}} or \code{\link[=dynplot_onedim]{dynplot_onedim()}}
}
