% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynplot_dimred.R
\name{dynplot_dimred}
\alias{dynplot_dimred}
\title{Plot a trajectory based on a dimensionality reduction}
\usage{
dynplot_dimred(dataset, trajectory = dataset, dimred = dataset$dimred)
}
\arguments{
\item{dataset}{A dynwrap dataset object, typically containing a trajectory.}

\item{trajectory}{The trajectory dynwrap object, if available.}

\item{dimred}{A dimensionality reduction matrix of the cells. Default is \code{dataset$dimred}.}
}
\value{
A ggplot2 object, with the processed data in \code{plot$data} and \code{attr(plot$data, "data")}.
}
\description{
Plot a trajectory based on a dimensionality reduction
}
\examples{
library(ggplot2)
data(example_bifurcating)
dataset <- example_bifurcating

dynplot_dimred(dataset) +
  geom_trajectory_segments() +
  geom_cell_point(size = 2, colour = "black") +
  geom_cell_point(aes(colour = milestone_percentages), size = 1.8) +
  geom_milestone_label(aes(fill = milestone_id)) +
  scale_milestones_color() +
  scale_milestones_fill()

dynplot_dimred(dataset) +
  geom_cell_point(aes(colour = select_feature_expression("G1", d = .data))) +
  scale_expression_color() +
  new_scale_color() +
  geom_trajectory_segments(aes(colour = milestone_percentages), size = 2) +
  geom_milestone_label(aes(fill = milestone_id)) +
  scale_milestones_color() +
  scale_milestones_fill()
}
\seealso{
\code{\link[=dynplot_dendro]{dynplot_dendro()}}, \code{\link[=dynplot_dimred]{dynplot_dimred()}}, \code{\link[=dynplot_graph]{dynplot_graph()}} or \code{\link[=dynplot_onedim]{dynplot_onedim()}}
}
