% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynplot_graph.R
\name{dynplot_graph}
\alias{dynplot_graph}
\title{Plot a trajectory as a graph}
\usage{
dynplot_graph(dataset, trajectory = dataset)
}
\arguments{
\item{dataset}{A dynwrap dataset object, typically containing a trajectory.}

\item{trajectory}{The trajectory dynwrap object, if available.}
}
\value{
A ggplot2 object, with the processed data in \code{plot$data} and \code{attr(plot$data, "data")}.
}
\description{
Plot a trajectory as a graph
}
\examples{
library(ggplot2)
data(example_bifurcating)
dataset <- example_bifurcating

dynplot_graph(dataset) +
  geom_trajectory_divergence() +
  geom_trajectory_segments(size = 2, color = "#333333", arrow_size = 0.5) +
  geom_cell_point(colour = "black", size = 3) +
  geom_cell_point(aes(colour = milestone_percentages), size = 2.8) +
  geom_milestone_label(aes(fill = milestone_id)) +
  scale_milestones_fill() +
  scale_milestones_color()
}
\seealso{
\code{\link[=dynplot_dendro]{dynplot_dendro()}}, \code{\link[=dynplot_dimred]{dynplot_dimred()}}, \code{\link[=dynplot_graph]{dynplot_graph()}} or \code{\link[=dynplot_onedim]{dynplot_onedim()}}
}
