% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynplot_onedim.R
\name{dynplot_onedim}
\alias{dynplot_onedim}
\title{Plot a trajectory as a graph}
\usage{
dynplot_onedim(
  dataset,
  trajectory = dataset,
  margin = 0.02,
  equal_cell_width = TRUE
)
}
\arguments{
\item{dataset}{A dynwrap dataset object, typically containing a trajectory.}

\item{trajectory}{The trajectory dynwrap object, if available.}
}
\value{
A ggplot2 object, with the processed data in \code{plot$data} and \code{attr(plot$data, "data")}.
}
\description{
Plot a trajectory as a graph
}
\examples{
data(example_bifurcating)
dataset <- example_bifurcating

dynplot_onedim(dataset, equal_cell_width = TRUE, margin = 0.05) +
  geom_trajectory_segments() +
  geom_trajectory_connection() +
  geom_cell_point(aes(colour = select_feature_expression("G2", d = .data))) +
  scale_expression_colour() +
  new_scale_colour() +
  geom_milestone_label(aes(y = -0.1, hjust = as.integer(type == "end")))

dynplot_onedim(dataset, equal_cell_width = TRUE, margin = 0.05) +
  geom_trajectory_segments(aes(color = milestone_percentages), size = 1, color = "#333333") +
  geom_trajectory_connection() +
  scale_milestones_colour() +
  new_scale_colour() +
  geom_cell_point(aes(colour = select_feature_expression("G2", d = .data))) +
  scale_expression_colour() +
  new_scale_colour() +
  geom_milestone_label(aes(y = -0.1, fill = milestone_id, hjust = as.integer(type == "end"))) +
  scale_milestones_fill()
}
\seealso{
\code{\link[=dynplot_dendro]{dynplot_dendro()}}, \code{\link[=dynplot_dimred]{dynplot_dimred()}}, \code{\link[=dynplot_graph]{dynplot_graph()}} or \code{\link[=dynplot_onedim]{dynplot_onedim()}}
}
