% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_velocity_arrow.R, R/geom_velocity_stream.R
\name{embed_arrows_cells}
\alias{embed_arrows_cells}
\alias{embed_arrows_random}
\alias{embed_arrows_grid}
\alias{embed_arrows_stream}
\title{Embed velocity arrows in a grid}
\usage{
embed_arrows_cells(cell_positions)

embed_arrows_random(cell_positions, n_cells = 100)

embed_arrows_grid(
  cell_positions,
  grid_n = c(15, 15),
  grid_bandwidth = 1/3,
  max_arrow_length = NULL,
  filter = rlang::quo(mass > max(mass) * 0.1)
)

embed_arrows_stream(
  cell_positions,
  grid_bandwidth = 1/3,
  filter = rlang::quo(mass > max(mass) * 0.1)
)
}
\arguments{
\item{cell_positions}{The dimensionality reduction which contains at least x, y, x_future and y_future}

\item{grid_n}{Number of rows and columns in the grid}

\item{grid_bandwidth}{Bandwidth relative to the spacing of the grid}
}
\description{
Embed velocity arrows in a grid
}
