% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_data_.R
\name{facet_grid_data}
\alias{facet_grid_data}
\title{Grid facetting based on data}
\usage{
facet_grid_data(...)
}
\description{
Grid facetting based on data
}
\examples{
library(ggplot2)
cutoff <- function(x, cutoff) {
  x > cutoff
}

d <- data.frame(
  x = 1:100,
  y = sqrt(1:100)
)

ggplot(d) +
  geom_text(aes(x, y, color = x > cutoff, label = hello)) +
  facet_grid_data(hello = c("a", "b", "c"), cutoff = c(30, -100, 10))

}
