% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cell_.R, R/geom_expression.R
\name{geom_cell_point}
\alias{geom_cell_point}
\alias{geom_cell_hex}
\alias{geom_expression_tile}
\title{Plotting cells}
\usage{
geom_cell_point(mapping = NULL, data = construct_get_cell_info(), ...,
  show.legend = NA)

geom_cell_hex(mapping = NULL, data = construct_get_cell_info(),
  bins = 100, ..., show.legend = NA)

geom_expression_tile(mapping = NULL, ..., show.legend = NA,
  rescale = dynutils::scale_quantile,
  data = construct_get_expression_info(rescale = rescale))
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes().}

\item{data}{A function created by \code{\link[=construct_get_cell_info]{construct_get_cell_info()}}.}

\item{show.legend}{Whether to show a legend for this geom}

\item{mapping}{Set of aesthetic mappings created by aes().}

\item{data}{A function created by \code{\link[=construct_get_cell_info]{construct_get_cell_info()}}.}

\item{show.legend}{Whether to show a legend for this geom}
}
\description{
Plotting cells

Plotting expression
}
