% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_cell_contour.R
\name{geom_cell_contour}
\alias{geom_cell_contour}
\title{Plot contour around cells based}
\usage{
geom_cell_contour(
  mapping = NULL,
  data = construct_get_cell_info(),
  relative_density_cutoff = 0.2,
  relative_bandwidth = 0.2,
  padding = 1,
  resolution = 200,
  ...,
  show.legend = NA
)
}
\arguments{
\item{relative_density_cutoff}{At whtat level of density the contour should be drawn, should be between 0 and 1}

\item{relative_bandwidth}{Bandwidth calculated relative to the x and y limits of the points, should be between 0 and 1}

\item{padding}{How much padding to add to the limits, to avoid the contour to be drawn outside the plot}

\item{resolution}{The higher, the more accurate the polygon will be drawn at the cost of longer computing/drawing time}
}
\description{
Plot contour around cells based
}
