% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_feature_.R
\name{geom_feature_text}
\alias{geom_feature_text}
\title{Plotting feature text}
\usage{
geom_feature_text(mapping = NULL, data = construct_get_feature_info(),
  ..., show.legend = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by aes().}

\item{data}{A function created by \code{\link[=get_feature_info_constructor]{get_feature_info_constructor()}}.}

\item{show.legend}{Whether to show a legend for this geom}
}
\description{
Plotting feature text
}
