% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_milestone.R
\name{define_milestone_colors}
\alias{define_milestone_colors}
\alias{color_milestone_percentages}
\alias{milestone_percentage_breaks}
\title{Helper functions for coloring milestones}
\usage{
define_milestone_colors(milestone_colors, milestone_ids)

color_milestone_percentages(milestone_percentages, milestone_colors)

milestone_percentage_breaks(milestone_ids)
}
\arguments{
\item{milestone_colors}{The matrix linking milestones to RGB, as created by define_milestone_colors}

\item{milestone_percentages}{A tibble of milestone percentages of a particular cell}
}
\description{
Helper functions for coloring milestones
}
