% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_new_scale.R
\name{new_scale}
\alias{new_scale}
\alias{new_scale_color}
\alias{new_scale_colour}
\alias{new_scale_fill}
\title{Add a new scale}
\usage{
new_scale(new_aes)

new_scale_color()

new_scale_colour()

new_scale_fill()
}
\description{
Add a new scale
}
