% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_expression.R
\name{scale_expression_colour}
\alias{scale_expression_colour}
\alias{scale_expression_color}
\alias{scale_expression_fill}
\title{Scale expression values}
\usage{
scale_expression_colour(name = "Expression", ...)

scale_expression_color(name = "Expression", ...)

scale_expression_fill(name = "Expression", ...)
}
\description{
Scale expression values
}
