% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynplot.R
\name{dynplot}
\alias{dynplot}
\title{Create a dynplot with a specified layout}
\usage{
dynplot(dataset, trajectory, layout)
}
\arguments{
\item{dataset}{A dynwrap dataset object, typically containing a trajectory.}

\item{trajectory}{The trajectory dynwrap object, if available.}

\item{layout}{A layout list object. Can contain data frames \code{cell_positions}, \code{milestone_positions}, \code{edge_positions}, \code{segment_positions}, \code{segment_progressions}.}
}
\value{
A ggplot2 object, with the processed data in \code{plot$data} and \code{attr(plot$data, "data")}.
}
\description{
It's recommended you use one of the layout-specific dynplot functions such as
\code{\link[=dynplot_dendro]{dynplot_dendro()}}, \code{\link[=dynplot_dimred]{dynplot_dimred()}}, \code{\link[=dynplot_graph]{dynplot_graph()}} or \code{\link[=dynplot_onedim]{dynplot_onedim()}}
to start creating a new dynplot visualisation. Only use this function if you
know how dynplot layouts work.
}
\examples{
library(ggplot2)
dynplot_dimred(example_bifurcating) +
  geom_cell_point(aes(colour = select_feature_expression("G1", d = .data))) +
  scale_expression_colour() +
  new_scale_colour() +
  geom_trajectory_segments(aes(colour = edge_id))

}
\seealso{
\code{\link[=dynplot_dendro]{dynplot_dendro()}}, \code{\link[=dynplot_dimred]{dynplot_dimred()}}, \code{\link[=dynplot_graph]{dynplot_graph()}} or \code{\link[=dynplot_onedim]{dynplot_onedim()}}
}
