% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_dendro.R
\name{dynplot_dendro}
\alias{dynplot_dendro}
\title{Dendrogram layout of a trajectory}
\usage{
dynplot_dendro(
  dataset,
  trajectory = dataset,
  diag_offset = 0.05,
  y_offset = 0.2
)
}
\arguments{
\item{dataset}{A dataset created by \code{\link[dynwrap:wrap_data]{wrap_data()}} or \code{\link[dynwrap:wrap_expression]{wrap_expression()}}}

\item{trajectory}{The trajectory as created by \code{\link[dynwrap:infer_trajectory]{infer_trajectory()}} or \code{\link[dynwrap:add_trajectory]{add_trajectory()}}}

\item{diag_offset}{The x-offset (percentage of the edge lenghts) between milestones}
}
\description{
Dendrogram layout of a trajectory
}
\keyword{plot_trajectory}
