% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_data_.R
\name{facet_wrap_data}
\alias{facet_wrap_data}
\title{Wrap facetting based on data}
\usage{
facet_wrap_data(..., nrow = NULL, ncol = 4, labeller = NULL)
}
\description{
Wrap facetting based on data
}
\examples{
library(ggplot2)
cutoff <- function(x, cutoff) {
  x > cutoff
}

d <- data.frame(
  x = 1:100,
  y = sqrt(1:100)
)

ggplot(d) +
  geom_text(aes(x, y, color = x > cutoff, label = hello)) +
  facet_wrap_data(cutoff = c(1, 5, 15))

}
